KG dataloader updation
