+++ vpeleaderboard/data/__init__.py
@@ -2,4 +2,4 @@
 Data module containing code and utilities.
 """
 
-__all__ = ["src", "models","utils"]
+__all__ = ["src","utils"]
