+++ vpeleaderboard/data/src/__init__.py
@@ -2,7 +2,5 @@
 """
 Code module containing SBML processing scripts.
 """
-from . import basico_model
-from . import sys_bio_model
-from . import sbml_dataloader
-from . import kg
+from . import SBML
+from . import KG
