+++ vpeleaderboard/data/src/kg/__init__.py
@@ -1,6 +1,4 @@
 from . import biobridge_datamodule_hetero
-from . import primekg
 from . import biobridge_primekg
-from . import dataset
 
-__all__ = ["biobridge_datamodule_hetero", "primekg", "biobridge_primekg", "dataset"]
+__all__ = ["biobridge_datamodule_hetero", "biobridge_primekg"]
