+++ vpeleaderboard/data/src/kg/biobridge_primekg.py
@@ -1,5 +1,5 @@
 """
-Class for loading BioBridgePrimeKG dataset.
+Class for loading BioBridgePrimeKG dataset and  PrimeKG nodes and edges data set.
 """
 
 import os
@@ -9,8 +9,216 @@ import requests
 import numpy as np
 import pandas as pd
 from tqdm import tqdm
-from .dataset import Dataset
-from .primekg import PrimeKG
+from abc import ABC, abstractmethod
+
+class Dataset(ABC):
+    """
+    Abstract class for dataset.
+    """
+    @abstractmethod
+    def setup(self):
+        """
+        A method to set up the dataset.
+        """
+
+    @abstractmethod
+    def load_data(self):
+        """
+        A method to load the dataset and potentially preprocess it.
+        """
+
+class PrimeKG(Dataset):
+    """
+    Class for loading PrimeKG dataset.
+    It downloads the data from the Harvard Dataverse and stores it in the local directory.
+    The data is then loaded into pandas DataFrame of nodes and edges.
+    """
+
+    def __init__(self, local_dir: str = "../../../../data/primekg/"):
+        """
+        Constructor for PrimeKG class.
+
+        Args:
+            local_dir (str): The local directory where the data will be stored.
+        """
+        self.name: str = "primekg"
+        self.server_path: str = "https://dataverse.harvard.edu/api/access/datafile/"
+        self.file_ids: dict = {"nodes": 6180617, "edges": 6180616}
+        self.local_dir: str = local_dir
+
+        # Attributes to store the data
+        self.nodes: pd.DataFrame = None
+        self.edges: pd.DataFrame = None
+
+        # Set up the dataset
+        self.setup()
+
+    def setup(self):
+        """
+        A method to set up the dataset.
+        """
+        # Make the directory if it doesn't exist
+        os.makedirs(os.path.dirname(self.local_dir), exist_ok=True)
+
+
+    def _download_file(self, remote_url:str, local_path: str):
+        """
+        A helper function to download a file from remote URL to the local directory.
+
+        Args:
+            remote_url (str): The remote URL of the file to be downloaded.
+            local_path (str): The local path where the file will be saved.
+        """
+        response = requests.get(remote_url, stream=True, timeout=300)
+        response.raise_for_status()
+        progress_bar = tqdm(
+            total=int(response.headers.get("content-length", 0)),
+            unit="iB",
+            unit_scale=True,
+        )
+        with open(local_path, "wb") as file:
+            for data in response.iter_content(1024):
+                progress_bar.update(len(data))
+                file.write(data)
+        progress_bar.close()
+
+    def _load_nodes(self) -> pd.DataFrame:
+        """
+        Private method to load the nodes dataframe of PrimeKG dataset.
+        This method downloads the nodes file from the Harvard Dataverse if it does not exist
+        in the local directory. Otherwise, it loads the data from the local directory.
+        It further processes the dataframe of nodes and returns it.
+
+        Returns:
+            The nodes dataframe of PrimeKG dataset.
+        """
+        local_file = os.path.join(self.local_dir, f"{self.name}_nodes.tsv.gz")
+        if os.path.exists(local_file):
+            print(f"{local_file} already exists. Loading the data from the local directory.")
+
+            # Load the dataframe from the local directory and assign it to the nodes attribute
+            nodes = pd.read_csv(local_file, sep="\t", compression="gzip", low_memory=False)
+        else:
+            print(f"Downloading node file from {self.server_path}{self.file_ids['nodes']}")
+
+            # Download the file from the Harvard Dataverse with designated file_id for node
+            self._download_file(f"{self.server_path}{self.file_ids['nodes']}",
+                                os.path.join(self.local_dir, "nodes.tab"))
+
+            # Load the downloaded file into a pandas DataFrame
+            nodes = pd.read_csv(os.path.join(self.local_dir, "nodes.tab"),
+                                     sep="\t", low_memory=False)
+
+            # Further processing of the dataframe
+            nodes = nodes[
+                ["node_index", "node_name", "node_source", "node_id", "node_type"]
+            ]
+
+            # Store compressed dataframe in the local directory
+            nodes.to_csv(local_file, index=False, sep="\t", compression="gzip")
+
+        return nodes
+
+    def _load_edges(self, nodes: pd.DataFrame) -> pd.DataFrame:
+        """
+        Private method to load the edges dataframe of PrimeKG dataset.
+        This method downloads the edges file from the Harvard Dataverse if it does not exist
+        in the local directory. Otherwise, it loads the data from the local directory.
+        It further processes the dataframe of edges and returns it.
+
+        Args:
+            nodes (pd.DataFrame): The nodes dataframe of PrimeKG dataset.
+
+        Returns:
+            The edges dataframe of PrimeKG dataset.
+        """
+        local_file = os.path.join(self.local_dir, f"{self.name}_edges.tsv.gz")
+        if os.path.exists(local_file):
+            print(f"{local_file} already exists. Loading the data from the local directory.")
+
+            # Load the dataframe from the local directory and assign it to the edges attribute
+            edges = pd.read_csv(local_file, sep="\t", compression="gzip", low_memory=False)
+        else:
+            print(f"Downloading edge file from {self.server_path}{self.file_ids['edges']}")
+
+            # Download the file from the Harvard Dataverse with designated file_id for edge
+            self._download_file(f"{self.server_path}{self.file_ids['edges']}",
+                                os.path.join(self.local_dir, "edges.csv"))
+
+            # Load the downloaded file into a pandas DataFrame
+            edges = pd.read_csv(os.path.join(self.local_dir, "edges.csv"),
+                                     sep=",", low_memory=False)
+
+            # Further processing of the dataframe
+            edges = edges.merge(
+                nodes, left_on="x_index", right_on="node_index"
+            )
+            edges.drop(["x_index"], axis=1, inplace=True)
+            edges.rename(
+                columns={
+                    "node_index": "head_index",
+                    "node_name": "head_name",
+                    "node_source": "head_source",
+                    "node_id": "head_id",
+                    "node_type": "head_type",
+                },
+                inplace=True,
+            )
+            edges = edges.merge(
+                nodes, left_on="y_index", right_on="node_index"
+            )
+            edges.drop(["y_index"], axis=1, inplace=True)
+            edges.rename(
+                columns={
+                    "node_index": "tail_index",
+                    "node_name": "tail_name",
+                    "node_source": "tail_source",
+                    "node_id": "tail_id",
+                    "node_type": "tail_type"
+                },
+                inplace=True,
+            )
+            edges = edges[
+                [
+                    "head_index", "head_name", "head_source", "head_id", "head_type",
+                    "tail_index", "tail_name", "tail_source", "tail_id", "tail_type",
+                    "display_relation", "relation",
+                ]
+            ]
+
+            # Store compressed dataframe in the local directory
+            edges.to_csv(local_file, index=False, sep="\t", compression="gzip")
+
+        return edges
+
+    def load_data(self):
+        """
+        Load the PrimeKG dataset into pandas DataFrame of nodes and edges.
+        """
+        print("Loading nodes of PrimeKG dataset ...")
+        self.nodes = self._load_nodes()
+
+        print("Loading edges of PrimeKG dataset ...")
+        self.edges = self._load_edges(self.nodes)
+
+    def get_nodes(self) -> pd.DataFrame:
+        """
+        Get the nodes dataframe of PrimeKG dataset.
+
+        Returns:
+            The nodes dataframe of PrimeKG dataset.
+        """
+        return self.nodes
+
+    def get_edges(self) -> pd.DataFrame:
+        """
+        Get the edges dataframe of PrimeKG dataset.
+
+        Returns:
+            The edges dataframe of PrimeKG dataset.
+        """
+        return self.edges
+
 
 class BioBridgePrimeKG(Dataset):
     """
@@ -22,8 +230,8 @@ class BioBridgePrimeKG(Dataset):
     """
 
     def __init__(self,
-                 primekg_dir: str = "../../../data/primekg/",
-                 local_dir: str = "../../../data/biobridge_primekg/",
+                 primekg_dir: str = "../../../../data/primekg/",
+                 local_dir: str = "../../../../data/biobridge_primekg/",
                  random_seed: int=0,
                  n_neg_samples: int=5):
         """
@@ -54,7 +262,7 @@ class BioBridgePrimeKG(Dataset):
         # Attributes to store the data
         self.primekg = None
         self.primekg_triplets = None
-        self.primekg_triplets_negative = None
+        # self.primekg_triplets_negative = None
         self.data_config = None
         self.emb_dict = None
         self.df_train = None
@@ -71,7 +279,7 @@ class BioBridgePrimeKG(Dataset):
         A method to set up the dataset.
         """
         # Make the directories if it doesn't exist
-        os.makedirs(os.path.dirname(self.primekg_dir), exist_ok=True)
+        # os.makedirs(os.path.dirname(self.primekg_dir), exist_ok=True)
         os.makedirs(os.path.dirname(self.local_dir), exist_ok=True)
 
         # Set the random seed
@@ -360,206 +568,6 @@ class BioBridgePrimeKG(Dataset):
                             sep="\t", compression="gzip", index=False)
 
         return df_train, df_node_train, df_test, df_node_test, triplets
-
-    # def _negative_sampling(self,
-    #                        batch_df: pd.DataFrame,
-    #                        process_index: int,
-    #                        index_map: dict,
-    #                        node_train_dict: dict) -> pd.DataFrame:
-    #     """
-    #     A helper function to perform negative sampling for a batch of triplets.
-    #     """
-    #     negative_y_index_list = []
-    #     for _, row in tqdm(batch_df.iterrows(),
-    #                        total=batch_df.shape[0],
-    #                        desc=f"Process {process_index}"):
-    #         x_index = row['head_index']
-    #         # y_index = row['y_index']
-    #         y_index_type = row['tail_type']
-    #         paired_y_index_list = index_map[x_index]
-
-    #         # sample a list of negative y_index
-    #         node_train_sub = node_train_dict[y_index_type]
-    #         negative_y_index = node_train_sub[
-    #             ~node_train_sub['node_index'].isin(paired_y_index_list)
-    #         ]['node_index'].sample(self.n_neg_samples).tolist()
-    #         negative_y_index_list.append(negative_y_index)
-
-    #     batch_df.loc[:, 'negative_tail_index'] = negative_y_index_list
-    #     return batch_df
-
-    # def _build_negative_triplets(self,
-    #                              chunk_size: int=100000,
-    #                              n_neg_samples: int=10):
-    #     """
-    #     Build the negative triplets for BioBridgePrimeKG dataset.
-    #     """
-    #     processed_file_path = os.path.join(self.local_dir,
-    #                                        "processed",
-    #                                        "triplet_train_negative.tsv.gz")
-    #     if os.path.exists(processed_file_path):
-    #         # Load the negative triplets from the local directory
-    #         with open(processed_file_path, "rb") as f:
-    #             triplets_negative = pd.read_csv(f, sep="\t", compression="gzip", low_memory=False)
-    #     else:
-    #         # Set the number samples for negative sampling
-    #         self.n_neg_samples = n_neg_samples
-
-    #         # Split node list by type
-    #         node_train_dict = {}
-    #         type_list = self.df_node_train['node_type'].unique()
-    #         for node_type in type_list:
-    #             node_train_dict[node_type] = self.df_node_train[
-    #                 self.df_node_train['node_type'] == node_type
-    #             ].reset_index(drop=True)
-
-    #         # create an index mapping from x_index to y_index
-    #         index_map = self.df_train[
-    #             ['head_index', 'tail_index']
-    #         ].drop_duplicates().groupby('head_index').agg(list).to_dict()['tail_index']
-
-    #         # Negative sampling
-    #         batch_df_list = []
-    #         for i in tqdm(range(0, self.df_train.shape[0], chunk_size)):
-    #             batch_df_list.append(self.df_train.iloc[i:i+chunk_size])
-    #         # Process negative sampling
-    #         results = [
-    #             self._negative_sampling(batch_df,
-    #                                     num_piece,
-    #                                     index_map,
-    #                                     node_train_dict)
-    #                                     for num_piece, batch_df in enumerate(batch_df_list)
-    #         ]
-
-    #         # Store the negative triplets
-    #         triplets_negative = pd.concat(results, axis=0)
-    #         triplets_negative.to_csv(processed_file_path,
-    #                                  sep="\t", compression="gzip", index=False)
-
-    #     # Set attribute
-    #     self.primekg_triplets_negative = triplets_negative
-
-    #     return triplets_negative
-
-    # def load_data(self,
-    #               build_neg_triplest: bool= False,
-    #               chunk_size: int=100000,
-    #               n_neg_samples: int=10):
-
-    def load_data(self):
-        """
-        Load the BioBridgePrimeKG dataset into pandas DataFrame of nodes and edges.
-
-        Args:
-            build_neg_triplest (bool): Whether to build negative triplets.
-            chunk_size (int): The chunk size for negative sampling.
-            n_neg_samples (int): The number of negative samples for negative sampling.
-        """
-        # Load PrimeKG dataset
-        print("Loading PrimeKG dataset...")
-        self.primekg = self._load_primekg()
-
-        # Load data config file of BioBridgePrimeKG
-        print("Loading data config file of BioBridgePrimeKG...")
-        self.data_config = self._load_data_config()
-
-        # Build node embeddings
-        print("Building node embeddings...")
-        self.emb_dict = self._build_node_embeddings()
-
-        # Build full triplets
-        print("Building full triplets...")
-        self.primekg_triplets, self.node_info_dict = self._build_full_triplets()
-
-        # Build train-test split
-        print("Building train-test split...")
-        self.df_train, self.df_node_train, self.df_test, self.df_node_test, self.primekg_triplets =\
-        self._build_train_test_split()
-
-        # if build_neg_triplest:
-        #     # Build negative triplets
-        #     print("Building negative triplets...")
-        #     self.primekg_triplets_negative = self._build_negative_triplets(
-        #         chunk_size=chunk_size,
-        #         n_neg_samples=n_neg_samples
-        #     )
-
-    def set_random_seed(self, seed: int):
-        """
-        Set the random seed for reproducibility.
-
-        Args:
-            seed (int): The random seed value.
-        """
-        np.random.seed(seed)
-
-    def get_primekg(self) -> PrimeKG:
-        """
-        Get the PrimeKG dataset.
-
-        Returns:
-            The PrimeKG dataset.
-        """
-        return self.primekg
-
-    def get_data_config(self) -> dict:
-        """
-        Get the data config file of BioBridgePrimeKG dataset.
-
-        Returns:
-            The data config file of BioBridgePrimeKG dataset.
-        """
-        return self.data_config
-
-    def get_node_embeddings(self) -> dict:
-        """
-        Get the node embeddings for BioBridgePrimeKG dataset.
-
-        Returns:
-            The dictionary of node embeddings.
-        """
-        return self.emb_dict
-
-    def get_primekg_triplets(self) -> pd.DataFrame:
-        """
-        Get the full triplets for BioBridgePrimeKG dataset.
-
-        Returns:
-            The full triplets for BioBridgePrimeKG dataset.
-        """
-        return self.primekg_triplets
-
-    # def get_primekg_triplets_negative(self) -> pd.DataFrame:
-    #     """
-    #     Get the negative triplets for BioBridgePrimeKG dataset.
-
-    #     Returns:
-    #         The negative triplets for BioBridgePrimeKG dataset.
-    #     """
-    #     return self.primekg_triplets_negative
-
-    def get_train_test_split(self) -> dict:
-        """
-        Get the train-test split for BioBridgePrimeKG dataset.
-
-        Returns:
-            The train-test split for BioBridgePrimeKG dataset.
-        """
-        return {
-            "train": self.df_train,
-            "node_train": self.df_node_train,
-            "test": self.df_test,
-            "node_test": self.df_node_test
-        }
-
-    def get_node_info_dict(self) -> dict:
-        """
-        Get the node information dictionary for BioBridgePrimeKG dataset.
-
-        Returns:
-            The node information dictionary for BioBridgePrimeKG dataset.
-        """
-        return self.node_info_dict
 
     # def _negative_sampling(self,
     #                        batch_df: pd.DataFrame,
